requirements.txt
Django==4.1.7
django_allauth==0.52.0
django_rest_auth==0.9.5
djangorestframework==3.14.0
protobuf==4.22.0
requests==2.28.2
